+++ lib/rspec/xml_helpers/xml_matcher.rb
@@ -9,19 +9,36 @@ module RSpec
     # @api private
     class XmlMatcher
 
-      EXPECTED_XML = 'Expected XML, but encountered an error while ' \
-        'parsing the actual value: %<error>s'
+      BAD_EXPECTED = <<~MSG
+        Encountered an error parsing the match_xml expected value:
+          %<error>s
+
+        Invalid XML:
+          %<got>s
+      MSG
+
+      BAD_ACTUAL = <<~MSG
+        Expected actual to be an XML string, encountered error while parsing:
+          %<error>s
+
+        Invalid XML:
+          %<got>s
+      MSG
 
       def initialize(expected, colorize: true)
         @expected = XmlHelpers.normalize_xml(expected)
         @diff_format = colorize ? :color : :text
+      rescue NawsXml::ParseError => error
+        raise ArgumentError, format(BAD_EXPECTED,
+          error: error.message, got: got(expected))
       end
 
       def matches?(actual)
         @actual = XmlHelpers.normalize_xml(actual)
         @expected == @actual
       rescue NawsXml::ParseError => error
-        @error_message = format(EXPECTED_XML, error: error.message)
+        @error_message = format(BAD_ACTUAL,
+          error: error.message, got: got(actual))
         false
       end
 
@@ -44,6 +61,12 @@ module RSpec
       def diff
         diff = Diffy::Diff.new(@expected, @actual).to_s(@diff_format)
         diff.lines[0..-2].join
+      end
+
+      def got(value)
+        got = value.inspect unless value.is_a?(String)
+        got = xml.inspect if got == ''
+        got
       end
 
     end